# pkg/apiclient/api/openapi.yaml
openapi: 3.0.1
info:
  contact: {}
  description: Daytona Server API
  title: Daytona Server API
  version: v0.0.0-dev
servers:
- url: http://localhost:3986/
security:
- Bearer: []
paths:
...
  /project-config/prebuild:
    get:
      description: List prebuilds
      operationId: ListPrebuilds
      parameters:
      - description: Config name
        in: path
        name: configName
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            '*/*':
              schema:
                items:
                  $ref: '#/components/schemas/PrebuildDTO'
                type: array
          description: OK
      summary: List prebuilds
      tags:
      - prebuild
    put:
      description: Set prebuild
      operationId: SetPrebuild
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePrebuildDTO'
        description: Prebuild
        required: true
      responses:
        "201":
          content: {}
          description: Created
      summary: Set prebuild
      tags:
      - prebuild
      x-codegen-request-body-name: prebuild
  /project-config/prebuild/process-git-event:
    post:
      description: ProcessGitEvent
      operationId: ProcessGitEvent
      requestBody:
        content:
          '*/*':
            schema:
              type: object
        description: Webhook event
        required: true
      responses:
        "200":
          content: {}
          description: OK
      summary: ProcessGitEvent
      tags:
      - prebuild
      x-codegen-request-body-name: workspace
  /project-config/prebuild/{prebuildId}:
    get:
      description: Get prebuild
      operationId: GetPrebuild
      parameters:
      - description: Prebuild ID
        in: path
        name: prebuildId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PrebuildDTO'
          description: OK
      summary: Get prebuild
      tags:
      - prebuild
  /project-config/prebuild/{projectConfigName}/{prebuildId}:
    delete:
      description: Delete prebuild
      operationId: DeletePrebuild
      parameters:
      - description: Project config name
        in: path
        name: projectConfigName
        required: true
        schema:
          type: string
      - description: Prebuild ID
        in: path
        name: prebuildId
        required: true
        schema:
          type: string
      responses:
        "204":
          content: {}
          description: No Content
      summary: Delete prebuild
      tags:
      - prebuild
...
components:
  schemas:
...
    CreatePrebuildDTO:
      example:
        projectConfigName: projectConfigName
        commitInterval: 0
        id: id
        branch: branch
        runAtInit: true
        triggerFiles:
        - triggerFiles
        - triggerFiles
      properties:
        branch:
          type: string
        commitInterval:
          type: integer
        id:
          type: string
        projectConfigName:
          type: string
        runAtInit:
          type: boolean
        triggerFiles:
          items:
            type: string
          type: array
      required:
      - branch
      - commitInterval
      - projectConfigName
      - runAtInit
      - triggerFiles
      type: object
...
    PrebuildDTO:
      example:
        projectConfigName: projectConfigName
        commitInterval: 0
        id: id
        branch: branch
        triggerFiles:
        - triggerFiles
        - triggerFiles
      properties:
        branch:
          type: string
        commitInterval:
          type: integer
        id:
          type: string
        projectConfigName:
          type: string
        triggerFiles:
          items:
            type: string
          type: array
      required:
      - branch
      - commitInterval
      - id
      - projectConfigName
      - triggerFiles
      type: object
...
  securitySchemes:
    Bearer:
      description: '"Type ''Bearer TOKEN'' to correctly set the API Key"'
      in: header
      name: Authorization
      type: apiKey
...